AUTHLIB-5 password reset
Complete password reset functionality.

Removed @ValidPassword annotation from User#password in favor of
manually validating the password. Moved the logic into a single method
that can be called when changing or resetting passwords.

There is still no logic to mark an LDAP user, but this should move us
closer in that direction.