AUTHLIB-41 ldap button fixes ...
Removed contextPath property from BaseSecurityConfiguration to fix startup issues without a context path set.

Made 'FilterBasedLdapUserSearch ldapSearch' field optional otherwise the application won't start if only table-based is enabled.

Reworked the LDAP User button. If LDAP only is enabled, only show the button. If LDAP and table-based are enabled, show the checkbox and disable the LDAP button by default - clicking the checkbox enables the LDAP button. This was the previous behavior.

Also add a label for the OHSU User checkbox and fix styles since it adds a bottom margin. Also fix the text, it wasn't using the text-danger style and it was too large and wrapping on the page. Added the small class.

Removed the mutedDiv function - wasn't being used.