SHIFT-186 fix email validation edge case
Michelle noticed that when she edited a users email address after they had logged in and set it for the first time, and the user tries to change their email address, they got a 500. Additionally the email notification is sent to the old incorrect address causing an additional exception.

This is related to the noemail profile. To allow for an optional email field on the User entity I had added a validation group. The field is only validated with the `Emailable` validation group. Effectively that field has never been validated when using the noemail profile. I changed the field to `type=email` which handles front-end validation, and added an additional check in the POST.